tests/tt_metal/tt_metal/data_movement/python/requirements.txt
pyyaml
loguru
matplotlib
numpy
pytest
